Modificacion de entidades: se corrigieron las relaciones entre tablas, se especifico el: @ManyToOne(fetch = FetchType.LAZY)
la migracion ya esta correcta y las entidades y relaciones se crearon correctamente en la base de datos

+++ AgroNet-B/src/main/java/com/tyrservices/agronetb/Models/entidades/CategoriaProducto.java
@@ -13,10 +13,11 @@ import lombok.ToString;
 @ToString
 @Table(name = "categoria_productos")
 public class CategoriaProducto {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "codigo_categoria")
-    private Integer codigoCategoria;
+    private Long codigoCategoria;
 
     @Column(name = "nombre_categoria", nullable = false, length = 100)
     private String nombreCategoria;
